+++ react_native/src/screens/ArticleDetail.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import {
+  SafeAreaView, Text,
+} from 'react-native';
+
+const ArticleDetail = () => {
+  return (
+    <SafeAreaView>
+      <Text>hello</Text>
+    </SafeAreaView>
+  );
+};
+
+export default ArticleDetail;
